Shared init migrate function
